==> FinBERT-LSTM-main/7_lstm_model_bert.py <==
import numpy as np 
import pandas as pd 
from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics import mean_absolute_error
from sklearn.metrics import mean_absolute_percentage_error
import tensorflow as tf

# hyperparameters
split = (0.85)
sequence_length = 10
epochs = 100
learning_rate = 0.02



# loading stock price and news data
stock_data = pd.read_csv("stock_price.csv")
news_data = pd.read_csv("sentiment.csv")
stock_column = ['Close']
news_column = ['FinBERT score']


len_stock_data = stock_data.shape[0]


# splitting data to train and test
train_examples = int(len_stock_data * split)
train = stock_data.get(stock_column).values[:train_examples]
train_sentiment = news_data.get(news_column).values[:train_examples]
test = stock_data.get(stock_column).values[train_examples:]
test_sentiment = news_data.get(news_column).values[train_examples:]
len_train = train.shape[0]
len_test = test.shape[0]
len_train_sentiment = train_sentiment.shape[0]
len_test_sentiment = test_sentiment.shape[0]


# normalizing data
scaler = MinMaxScaler()
train, test = scaler.fit_transform(train), scaler.fit_transform(test)


# splitting training data to x and y
X_train = []
for i in range(len_train - sequence_length):
    X_train.append(train[i : i + sequence_length])
len_X_train = len(X_train)
y_train = np.array(train[sequence_length:]).astype(float)


# splitting testing data to x and y
X_test = []
for i in range(len_test - sequence_length):
    X_test.append(test[i : i + sequence_length])
len_X_test = len(X_test)
y_test = np.array(test[sequence_length:]).astype(float)


print(f"X_train[0] = {X_train[10]}")
print("================")

# adding news sentiment to train and test
for i in range(len_X_train):
    X_train[i] = X_train[i].tolist()
    X_train[i].append(train_sentiment[sequence_length + i].tolist())
X_train = np.array(X_train).astype(float)

for i in range(len_X_test):
    X_test[i] = X_test[i].tolist()
    X_test[i].append(test_sentiment[sequence_length + i].tolist())
X_test = np.array(X_test).astype(float)


print(f"X_train shape = {X_train.shape}")
print(f"y_train shape = {y_train.shape}")
print(f"X_train[0] = {X_train[10]}")

# #creating LSTM model
# def model_create():
#     tf.random.set_seed(1234)
#     model = tf.keras.models.Sequential(
#         [
#             tf.keras.Input(shape = (X_train.shape[1], 1)),
#             tf.keras.layers.LSTM(units = 70, activation = "tanh", return_sequences = True),
#             tf.keras.layers.LSTM(units = 30, activation = "tanh", return_sequences = True),
#             tf.keras.layers.LSTM(units = 10, activation = "tanh", return_sequences = False),
#             tf.keras.layers.Dense(units = 1, activation = "linear")
#         ]
#     )

#     model.compile(
#         loss = tf.keras.losses.mean_squared_error,
#         optimizer = tf.keras.optimizers.Adam(learning_rate = learning_rate)
#     )

#     model.fit(
#         X_train, y_train,
#         epochs = epochs
#     )
#     return model


# # inverting normaliztion
# y_test = scaler.inverse_transform(y_test)



# # prediction on test set
# def predict(model):
#     predictions = model.predict(X_test)
#     predictions = scaler.inverse_transform(predictions.reshape(-1,1)).reshape(-1,1)
#     return predictions




# # evaluation
# def evaluate(predictions):
#     mae = mean_absolute_error(predictions, y_test)
#     mape = mean_absolute_percentage_error(predictions, y_test)
#     return mae, mape, (1 - mape)


# # trial runs
# def run_model(n):
#     total_mae = total_mape = total_acc = 0
#     for i in range(n):
#         model = model_create()
#         predictions = predict(model)
#         mae, mape, acc = evaluate(predictions)
#         total_mae += mae
#         total_mape += mape 
#         total_acc += acc 
#     return (total_mae / n), (total_mape / n), (total_acc / n)


# mae, mape, acc = run_model(1)

# print(f"Mean Absolute Error = {mae}")
# print(f"Mean Absolute Percentage Error = {mape}%")
# print(f"Accuracy = {acc}")
